### 2018-10-21 #1037 @whedon
[REVIEW]: GPUE: Graphics Processing Unit Gross-Pitaevskii Equation solver
**Submitting author:** @leios (<a href="http://orcid.org/0000-0002-3243-8918">James Schloss</a>)
**Repository:** <a href="https://github.com/GPUE-group/GPUE" target ="_blank">https://github.com/GPUE-group/GPUE</a>
**Version:** v1.0
