### 2018-10-21 #1037 @labarba
@mgalloy, @markbasham — Thank you for agreeing to review this submission for JOSS, we really appreciate your volunteer efforts to make this adventure in new-wave publishing a success!

This REVIEW issue thread is the main place for proceeding with the review, but you are encouraged to open new issues in the repository of the submission, as appropriate, and linking to them here, to keep track. Feel free to ask any questions you may have!

The [JOSS Review Process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) is described in our documentation site.
