### 2018-10-21 #1037 @labarba
@leios : I notice that some DOI links in your reference list are broken. Please fix your bibtex file, push, and re-compile using the command `@whedon generate pdf` in its own comment.
