### 2018-10-23 #1037 @leios
I believe the DOIs have been fixed. Working on the contribution guidelines soon.

Thanks!
