### 2018-11-01 #1037 @markbasham
Hi @labarba and @mgalloy Finally back from my conference and over jetlag, and I'm taking a look at this review.  I have had a look through the documentation, and Im happy with what I need to do, but how do i tick the checkboxes at the start of this ticket?
