### 2018-11-01 #1037 @mgalloy
@markbasham Just make sure you are logged into GitHub and you should be able to check the boxes.
