### 2018-11-03 #1037 @leios
@markbasham What errors are you getting? This is obviously something we should look into, but I think the build instructions in the documentation should work for most linux systems.
