### 2018-11-03 #1037 @mgalloy
I am also having problems on my Mac (unfortunately, the only system with an NVIDIA GPU right now). I can send a more detailed report this weekend.
