### 2018-11-04 GPUE-group/GPUE#14 @markbasham
Difficulties with the build process
Hi,

Im trying to build GPUE with the instructions provided, but i get the problems as shown below  I have tried to use the standard makefile, but I also get the same issues, and I cannot fix them up by adding the -std=gnu++0x arguments.

I do have other gcc compilers avalilable, but this is the standard RHEL6 one which is the system that i have access to a GPU card on, and if i go above 5 there is an error.

Hope this helps

Mark

> git clone https://github.com/GPUE-group/GPUE.git
Initialized empty Git repository in /dls/tmp/ssg37927/test/GPUE/.git/
remote: Enumerating objects: 66, done.
remote: Total 2705 (delta 33), reused 45 (delta 20), pack-reused 2639
Receiving objects: 100% (2705/2705), 3.83 MiB | 650 KiB/s, done.
Resolving deltas: 100% (2057/2057), done.

> [ssg37927@cs04r-sc-com14-09 test]$ module load cuda/8.0

> [ssg37927@cs04r-sc-com14-09 test]$ nvcc --version
nvcc: NVIDIA (R) Cuda compiler driver
Copyright (c) 2005-2016 NVIDIA Corporation
Built on Tue_Jan_10_13:22:03_CST_2017
Cuda compilation tools, release 8.0, V8.0.61

> [ssg37927@cs04r-sc-com14-09 test]$ gcc --version
gcc (GCC) 4.4.7 20120313 (Red Hat 4.4.7-23)
Copyright (C) 2010 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.

> [ssg37927@cs04r-sc-com14-09 test]$ module load cmake

> [ssg37927@cs04r-sc-com14-09 test]$ cmake --version
cmake version 3.12.0
CMake suite maintained and supported by Kitware (kitware.com/cmake).

> [ssg37927@cs04r-sc-com14-09 test]$ cd GPUE/

> [ssg37927@cs04r-sc-com14-09 GPUE]$ cmake .
-- The CXX compiler identification is GNU 4.4.7
-- The CUDA compiler identification is NVIDIA 8.0.61
-- Check for working CXX compiler: /usr/bin/c++
-- Check for working CXX compiler: /usr/bin/c++ -- works
-- Detecting CXX compiler ABI info
-- Detecting CXX compile features
-- Check for working CUDA compiler: /dls_sw/apps/cuda/8.0/bin/nvcc
-- Check for working CUDA compiler: /dls_sw/apps/cuda/8.0/bin/nvcc -- works
-- Detecting CUDA compiler ABI info
-- Detecting CUDA compiler ABI info - done
-- Looking for C++ include pthread.h
-- Looking for C++ include pthread.h - found
-- Looking for pthread_create
-- Looking for pthread_create - not found
-- Looking for pthread_create in pthreads
-- Looking for pthread_create in pthreads - not found
-- Looking for pthread_create in pthread
-- Looking for pthread_create in pthread - found
-- Found Threads: TRUE
-- Configuring done
-- Generating done
-- Build files have been written to: /dls/tmp/ssg37927/test/GPUE

> [ssg37927@cs04r-sc-com14-09 GPUE]$ make
Scanning dependencies of target unitTest
[  4%] Building CUDA object CMakeFiles/unitTest.dir/src/unit_test.cu.o
nvcc warning : The 'compute_20', 'sm_20', and 'sm_21' architectures are deprecated, and may be removed in a future release (Use -Wno-deprecated-gpu-targets to suppress warning).
In file included from /usr/lib/gcc/x86_64-redhat-linux/4.4.7/../../../../include/c++/4.4.7/unordered_map:35,
                 from /dls/tmp/ssg37927/test/GPUE/src/../include/ds.h:24,
                 from /dls/tmp/ssg37927/test/GPUE/src/unit_test.cu:1:
/usr/lib/gcc/x86_64-redhat-linux/4.4.7/../../../../include/c++/4.4.7/c++0x_warning.h:31:2: error: #error This file requires compiler and library support for the upcoming ISO C++ standard, C++0x. This support is currently experimental, and must be enabled with the -std=c++0x or -std=gnu++0x compiler options.
make[2]: *** [CMakeFiles/unitTest.dir/src/unit_test.cu.o] Error 1
make[1]: *** [CMakeFiles/unitTest.dir/all] Error 2
make: *** [all] Error 2

> [ssg37927@cs04r-sc-com14-09 GPUE]$ make
[  4%] Building CUDA object CMakeFiles/unitTest.dir/src/unit_test.cu.o
nvcc warning : The 'compute_20', 'sm_20', and 'sm_21' architectures are deprecated, and may be removed in a future release (Use -Wno-deprecated-gpu-targets to suppress warning).
In file included from /usr/lib/gcc/x86_64-redhat-linux/4.4.7/../../../../include/c++/4.4.7/unordered_map:35,
                 from /dls/tmp/ssg37927/test/GPUE/src/../include/ds.h:24,
                 from /dls/tmp/ssg37927/test/GPUE/src/unit_test.cu:1:
/usr/lib/gcc/x86_64-redhat-linux/4.4.7/../../../../include/c++/4.4.7/c++0x_warning.h:31:2: error: #error This file requires compiler and library support for the upcoming ISO C++ standard, C++0x. This support is currently experimental, and must be enabled with the -std=c++0x or -std=gnu++0x compiler options.
make[2]: *** [CMakeFiles/unitTest.dir/src/unit_test.cu.o] Error 1
make[1]: *** [CMakeFiles/unitTest.dir/all] Error 2
make: *** [all] Error 2
