### 2018-11-04 GPUE-group/GPUE#14 @leios
I was able to replicate the error with the cmake build; however, the makefile build should still work in this case.

`git checkout Makefile` and modify the lines to point to your cuda install and to use the correct architecture. Does this still provide an error?

We'll be working on the cmake stuff today. We do explicitly specify C++11, so it is weird that an error is coming up.
