### 2018-11-04 GPUE-group/GPUE#14 @markbasham
@leios thanks for the help, if I just use the makefile as specified I get this.  I have K80 gpus, so I think I have the GPU_ARCH correct.

> [ssg37927@cs04r-sc-com14-05 test]$ git clone https://github.com/GPUE-group/GPUE.git
Initialized empty Git repository in /dls/tmp/ssg37927/test/GPUE/.git/
remote: Enumerating objects: 66, done.
remote: Total 2705 (delta 33), reused 45 (delta 20), pack-reused 2639
Resolving deltas: 100% (2057/2057), done.

> [ssg37927@cs04r-sc-com14-05 test]$ cd GPUE/

> [ssg37927@cs04r-sc-com14-05 GPUE]$ module load cuda/8.0

> [ssg37927@cs04r-sc-com14-05 GPUE]$ nvcc --version
nvcc: NVIDIA (R) Cuda compiler driver
Copyright (c) 2005-2016 NVIDIA Corporation
Built on Tue_Jan_10_13:22:03_CST_2017
Cuda compilation tools, release 8.0, V8.0.61

> [ssg37927@cs04r-sc-com14-05 GPUE]$ gcc --version
gcc (GCC) 4.4.7 20120313 (Red Hat 4.4.7-23)
Copyright (C) 2010 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.

> [ssg37927@cs04r-sc-com14-05 GPUE]$ which nvcc
/dls_sw/apps/cuda/8.0/bin/nvcc

> [ssg37927@cs04r-sc-com14-05 GPUE]$ vim Makefile

> [ssg37927@cs04r-sc-com14-05 GPUE]$ head Makefile
CUDA_HOME = /dls_sw/apps/cuda/8.0
#CUDA_HOME = /apps/free/cuda/7.5.18/
#CUTT_DIR = cutt/lib
GPU_ARCH        = sm_37
OS:=    $(shell uname)
ifeq ($(OS),Darwin)
CUDA_LIB        = $(CUDA_HOME)/lib
CUDA_HEADER     = $(CUDA_HOME)/include
CC              = $(CUDA_HOME)/bin/nvcc -ccbin /usr/bin/clang --ptxas-options=-v#-save-temps
CFLAGS          = -g -std=c++11 -Wno-deprecated-gpu-targets

> [ssg37927@cs04r-sc-com14-05 GPUE]$ make
/dls_sw/apps/cuda/8.0/bin/nvcc --ptxas-options=-v --compiler-options -Wall  -c -o fileIO.o -I/dls_sw/apps/cuda/8.0/include  -g -O0 -std=c++11 -Xcompiler '-std=c++11' -Xcompiler '-fopenmp'  -L/dls_sw/apps/cuda/8.0/lib64  -lcufft -Xcompiler "-fopenmp" -arch=sm_37 src/fileIO.cu -dc
nvcc warning : The -std=c++11 flag is not supported with the configured host compiler. Flag will be ignored.
cc1plus: error: unrecognized command line option "-std=c++11"
make: *** [fileIO.o] Error 1
