### 2018-11-04 GPUE-group/GPUE#14 @mlxd
GCC 4.4.x is quite an old compiler, so I expect there to be problems with the C++11 requirements in the build process. Even if there is some compilation, the String ABI differences may cause problems between GCC 5.2-ish changes (though again, this is something I may be wrong with). It may be that we have to specify a minumum compiler version in this case, such as 4.8/4.9. Is there any way for you to may use of a newer GCC version? If not, we may need to consider given a recommendation to use Spack as an optional means to get a newer GCC version, though this is just an option.

Conda can also provide GCC 7.2 as an option, or we can opt for Clang as the host compiler too, though build-from-scaratch can be quite a long process.
