### 2018-11-04 GPUE-group/GPUE#14 @mlxd
Just as a note, most development I have used most recent version of GCC , or Clang as the host compilers. I am curious as to the errors seen with higher GCC versions. We have mostly used Centos 7, Ubuntu 16.04, or Arch Linux, which tend to have GCC 4.8, 6+, and 8+ respectively.
