### 2018-11-04 GPUE-group/GPUE#14 @markbasham
tell me about it :) RHEL6 is very old but very stable apparently.....

However it works fine with gcc 4.9.3, could that be added to the build requirements, then this can be closed :)

Cheers

Mark
