### 2018-11-04 GPUE-group/GPUE#14 @mlxd
I can certainly appreciate stability. Arch Linux on the other hand has sometimes caused me problems for looking at it the wrong way.

Awesome, thanks for the detailed logs, they were really helpful to pin things down. I'll make the changes to the documentation in the morning, and recommend GCC 4.9.3 as a minimum recommended version with RH releases.

I'll do some testing of this tomorrow too with GCC versions on Arch just to confirm everything, and maybe @leios would be fine with likewise on Centos? I'd be happy to detail the process and results with what works and what does not.

Thanks again for the help @markbasham
We will get the final approval from you after this testing before closing this.

P.S. I recall @mgalloy also stating build issues on a Mac, which we may also be able to take care of in this instance (hopefully). If not, we can open a separate issue for that.
