### 2018-11-04 #1037 @markbasham
@leios the code all works for me, as do the example processes etc.  One thing i need to look at is the performance of the code, now it uses 84% of my GPU when running, so its pretty well making use of the system, but i don't have much else to base this on.  I have had a look at the documentation of GPELab but i cant see any runtimes.  do you know what the runtime for a process comparable to your documented example is?
