### 2018-11-04 #1037 @leios
Thanks for submitting the building issue. I'll be working on that today.

For the benchmarks, I think running a simple test, [like this one](https://web.archive.org/web/20171120181431/https://peterwittek.com/gpe-comparison.html), would be a good start for simple split-op performance testing. That link was from an older build of GPUE, but we have not lost any functionality since then (I don't even think the flags have changed significantly from that build). It also provides GPElab scripts for the test.

As a note: the test there is for a simple harmonic oscillator in 2D. This is now a unit test that is run with `gpue -u`. A lot of the 2/3D vortex functionality is harder to benchmark. GPElab can do it, but the documentation is hard to come by.
