### 2018-11-05 #1037 @markbasham
@leios final thing, is there no DOI for the final thesis in the paper? I couldent find anything obvious, but there are often university repos now for thesis, a link would do.
