### 2018-11-05 #1037 @leios
I cannot seem to find a DOI, but there is a permanent link: http://hdl.handle.net/11250/2467021.

This can be added to the reference section, if desired.
