### 2018-11-05 GPUE-group/GPUE#14 @leios
This error came about due to a recent change to the `CMakeLists.txt` file for building out of tree. Basically, building the unit tests requires us to test the fileIO, which means we need to specify the directory of an input file. This meant that we needed special build instructions for that file and we forgot to specify c++11 for that.

This has been fixed for gcc4.8 as of the latest commit. I will see if I can find a gcc4.4 build around to try with too. I think we also have macs to try building on with nvidia gpu's somewhere in the lab, so I will give them a go too.
