### 2018-11-05 GPUE-group/GPUE#14 @mgalloy
On my Mojave (10.14) Mac, I was getting the following error on configure:

> nvcc fatal   : The version ('10.0') of the host compiler ('Apple clang') is not supported

I downgraded to Command Line Tools for Xcode 7.3 because of comments in some Apple forums. This let me configure and compile, but linking causes the following error:

> [ 95%] Linking CUDA device code CMakeFiles/gpue.dir/cmake_device_link.o
> nvcc fatal   : Unknown option 'Wl,-rpath,/usr/local/cuda/lib'

Attaching [CMakeCache.txt](https://github.com/GPUE-group/GPUE/files/2546806/CMakeCache.txt) file and [the log of the build process](https://github.com/GPUE-group/GPUE/files/2546805/make.log).
