### 2018-11-05 GPUE-group/GPUE#14 @leios
Ah, thanks! I just got this same error on an old mac I found in the lab. I'll see if we can figure this out.
