### 2018-11-06 GPUE-group/GPUE#14 @mlxd
I am pretty sure those 2012 iMacs had GTX675M (or similar) cards, which would be Kepler generation (ie sm_3x). I think it should still be fine to build and run on that (please feel free to correct me if I am wrong).
