### 2018-11-08 GPUE-group/GPUE#14 @mgalloy
I'm looking into using our supercomputer, there are nodes with Teslas which I believe I can use.
