### 2018-11-12 #1037 @leios
Thanks a bunch for the review @markbasham!

As for the current state of the review: There was some difficulty building the code on the latest MacOS release (more info found here: https://github.com/GPUE-group/GPUE/issues/14). This has been noted, but it doesn't seem like there is anything we can do at this stage until Apple and Nvidia developers start talking to each other again.
