### 2018-11-21 #1037 @mgalloy
Finally got access to the machine. Everything looks good. I have completed my review and have checked off all items on the checklist.
