### 2018-11-22 #1037 @labarba
@leios : I now have some editorial suggestions to polish your paper.

Page 1.
[par. 1]

- “many interesting physical phenomenon” >> phenomena
- “allow for new discoveries that directly mimic what can be seen in experiments and are thus highly valuable for fundamental research”>> hyperbole. If the simulations mimic experiments, it’s not really a “new discovery,” is it. Could you tone this down? E.g., what’s the difference between “valuable” and “highly valuable”?  You can say “Numerical simulations of BECs that directly mimic what can be seen in experiments are valuable in fundamental research in these areas.”
- “In practice, the dynamics of BEC systems can often be found” >> (declutter) The dynamics of BEC systems can be found …
- “Though there are many methods to solve the GPE, one of the most straightforward…” >> One of the most straightforward of many methods for solving the GPE… (and you get rid of the syntactic expletive “there are”)
- “however, there are no generalized software packages available…” >> (make new sentence and) “No generalized software packages are available…” (again, get rid of syntactic expletive)
- “Even so, there are several…” >> Packages do exist for simulating BECs with other methods…

[par. 2]
- “with an emphasis on” >> emphasizing
- “vortex turbulence in two dimensions” >> 2D turbulence is an oxymoron; perhaps say “vortex dynamics in two and three dimensions.”

Page 2.
- “highly configurable physical system parameters” >> what’s the difference between configurable and highly configurable? Perhaps delete “highly.”
- “GPUE additionally features a highly performant numerical solver implementation” >> GPUE’s high-performance numerical solver improves over other suites (Refs.)
