### 2018-11-23 GPUE-group/GPUE#14 @leios
I believe the current difficulties have been resolved. Please reopen if this is not the case.
