### 2018-11-23 #1037 @leios
I believe this version of the paper addresses all these points.
