### 2018-12-06 #1037 @leios
I made a few more smallscale changes to the paper and I believe it is good to go. Thanks for all the help so far and let me know what else needs to be done.
