### 2018-12-10 #1037 @labarba
This remains:

* “allow for new discoveries that directly mimic what can be seen in experiments and are thus highly valuable for fundamental research”>> hyperbole. If the simulations mimic experiments, it’s not really a “new discovery,” is it. Could you tone this down? E.g., what’s the difference between “valuable” and “highly valuable”? You can say “Numerical simulations of BECs that directly mimic what can be seen in experiments are valuable in fundamental research in these areas.”
