### 2018-12-10 #1037 @labarba
Actually, looks like several of my editorial suggestions were ignored. Can you go over the items in my comment above? https://github.com/openjournals/joss-reviews/issues/1037#issuecomment-440960650
