### 2018-12-10 #1037 @labarba
Typo: “No generalized software packages are vailable…” >> available
