### 2018-12-10 #1037 @labarba
"features a numerical solver with a greater performance than other available suites" >> This reads like an unsubstantiated performance claim. Is there justification somewhere for this? (in the tests, in the docs…)
