### 2018-12-10 #1037 @leios
There is a citation for performance tests after the latter claim (https://peterwittek.com/gpe-comparison.html). I suppose I could put the actual numbers in there and say that we are 206x faster than GPELab, 5.5x faster than TS-MPI CPU (the fastest known CPU implementation), and 1.19 faster than TS-MPI GPU in real-time and 147x faster than GPELab, 3.48x faster than TS-MPI CPU, and 1.21x faster than TS-MPI GPU in imaginary time; however, these are all in the reference.
