### 2018-12-10 #1037 @leios
It did. The previous paper said the following: `The above features enable highly configurable physical system parameters, and allow for the simulation of state-of-the-art system dynamics. GPUE additionally features a highly performant numerical solver implementation, with performance greater than other available suites.` By your request, this was modified to simply `GPUE additionally features a numerical solver with a greater performance than other available suites`
