### 2018-12-10 #1037 @labarba
My editorial suggestion above, in https://github.com/openjournals/joss-reviews/issues/1037#issuecomment-440960650

* “GPUE additionally features a highly performant numerical solver implementation” >> GPUE’s high-performance numerical solver improves over other suites (Refs.)

This mild claim didn't raise alarm bells. But your version:

> features a numerical solver with a greater performance than other available suites

… raised a question in my head: Is this a performance claim? Is it justified? Did the reviewers check that? Because the reviewers have an item to check off that says: "If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)"

I think that if you say "my software gives greater performance than others," readers will want evidence.  You can soften the language, or give a link or citation to evidence. I don't think you should add numbers in the paper now, because we'd have to go back to the reviewers
