### 2018-12-10 #1037 @leios
Alright, great! The link is there already, so I'll fix the typo and regenerate the paper?
