### 2018-12-10 #1037 @leios
> GPUE additionally features a numerical solver with a greater performance than other available suites

is now

> GPUE additionally features a numerical solver that improves on other available suites

and

> No generalized software packages are vailable

is now

> No software packages are available
