### 2018-12-11 #1037 @arfon
Fixed a small bug @labarba. This should be good to go now.
