### 2018-12-11 #1037 @arfon
Actually, I've broken this in a more complex way attempting a fix. Will process manually on my end until I can put a proper fix in place.
