### 2018-12-11 #1037 @arfon
@leios - your paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.01037 :zap::rocket::boom:
