### 2018-12-12 #1037 @labarba
One thing I forgot to say,  @leios — the author list on the Zenodo archive doesn't match the author list of the paper. Zenodo grabs authors from commit history: you have to edit manually to match!
