### 2018-12-13 #1037 @arfon
> Since there are only 2 authors, the et al. seems odd. It's not a big deal if it cannot be changed.

Sorry, this can't be easily changed.
